Add files via upload
The Final paper describes most the things in the project and how it works picture breaks down this complex process to a flowchart. 100X1 is the poster for this project and Smart Glasses.ino is the arduino file. The speech to text apk is the android application and the Speech to Text.aia is the code